Separated banking, energy and discovery
