+++ src/components/comparison/BankingComparisonPanel.js
@@ -14,16 +14,16 @@ import TableHead from '@material-ui/core/TableHead'
 import TableRow from '@material-ui/core/TableRow'
 import {productDataKeys} from '../../utils/dict'
 import {format} from '../../utils/datetime'
-import AdditionalInfo from '../data/AdditionalInfo'
+import AdditionalInfo from '../data/banking/AdditionalInfo'
 import ecomp from '../../utils/enum-comp'
-import Bundle from '../data/Bundle'
-import Constraint from '../data/Constraint'
-import DepositRate from '../data/DepositRate'
-import LendingRate from '../data/LendingRate'
-import Eligibility from '../data/Eligibility'
-import Feature from '../data/Feature'
-import Fee from '../data/Fee'
-import CardArt from '../data/CardArt'
+import Bundle from '../data/banking/Bundle'
+import Constraint from '../data/banking/Constraint'
+import DepositRate from '../data/banking/DepositRate'
+import LendingRate from '../data/banking/LendingRate'
+import Eligibility from '../data/banking/Eligibility'
+import Feature from '../data/banking/Feature'
+import Fee from '../data/banking/Fee'
+import CardArt from '../data/banking/CardArt'
 
 const useStyles = makeStyles(theme => ({
   panel: {
