+++ src/components/data-source/DataSource.js
@@ -16,7 +16,7 @@ import {
   modifyDataSourceUrl
 } from '../../store/data-source'
 import { clearSelection} from '../../store/selection'
-import { deleteData, clearData } from '../../store/data'
+import { deleteData, clearData } from '../../store/banking/data'
 import {connect} from 'react-redux'
 import isUrl from '../../utils/url'
 import Snackbar from '@material-ui/core/Snackbar'
