+++ src/components/data-source/DataSourcePanel.js
@@ -20,9 +20,8 @@ import Autocomplete from '@material-ui/lab/Autocomplete'
 import { fade } from '@material-ui/core/styles/colorManipulator'
 import { loadDataSource, addDataSource, syncDataSources } from '../../store/data-source'
 import { loadVersionInfo, saveVersionInfo, setVersionsEditable, setVersionsReadOnly } from '../../store/version-info'
-import { startRetrieveProductList, retrieveProductList } from '../../store/data'
+import { startRetrieveProductList, retrieveProductList, clearData } from '../../store/banking/data'
 import { clearSelection} from '../../store/selection'
-import { clearData } from '../../store/data'
 import { normalise } from '../../utils/url'
 import Fab from '@material-ui/core/Fab'
 import Tooltip from '@material-ui/core/Tooltip'
