+++ src/components/data/banking/FeeDiscount.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import DiscountEligibility from './DiscountEligibility'
 import {makeStyles} from '@material-ui/core'
-import {translateDiscountType} from '../../utils/dict'
-import Duration from './Duration'
-import ecomp from '../../utils/enum-comp'
+import {translateDiscountType} from '../../../utils/dict'
+import Duration from '../Duration'
+import ecomp from '../../../utils/enum-comp'
 
 const useStyles = makeStyles(() => ({
   sectionTitle: {
