+++ src/components/data/banking/Product.js
@@ -14,10 +14,10 @@ import Feature from './Feature'
 import Fee from './Fee'
 import CardArt from './CardArt'
 import Checkbox from '@material-ui/core/Checkbox'
-import {deselectProduct, selectProduct} from '../../store/selection'
-import DateTime from './DateTime'
+import {deselectProduct, selectProduct} from '../../../store/selection'
+import DateTime from '../DateTime'
 import AdditionalInfo from './AdditionalInfo'
-import ecomp from '../../utils/enum-comp'
+import ecomp from '../../../utils/enum-comp'
 
 const useStyles = makeStyles(() => ({
   root: {
