+++ src/components/data/banking/RateTier.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import RateCondition from './RateCondition'
 import RateSubTier from './RateSubTier'
-import {translateRateApplicationMethod, translateUnitOfMeasure} from '../../utils/dict'
+import {translateRateApplicationMethod, translateUnitOfMeasure} from '../../../utils/dict'
 import {makeStyles} from '@material-ui/core'
 
 const useStyles = makeStyles(() => ({
