+++ src/components/data/energy/EnergyPanel.js
