+++ src/store/banking/comparison/reducer.js
